+++ mediation/admob/native/MTRGAdmobNativeCustomEvent.h
@@ -1,0 +1,13 @@
+//
+//  MTRGAdmobNativeCustomEvent.h
+//  MyTargetMediationApp
+//
+//  Created by Andrey Seredkin on 13.03.17.
+//  Copyright © 2017 Mail.ru Group. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface MTRGAdmobNativeCustomEvent : NSObject
+
+@end